requirements.txt
onnxruntime
numpy
opencv-python
pyyaml
django
djangorestframework
pillow